requirements.txt
Flask
openai
requests
yfinance
stripe
google-api-python-client
google-auth
google-auth-oauthlib
